warning fixes

+++ libft.h
@@ -106,8 +106,8 @@ void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 
 void				*t_buff_init(t_buff *b, size_t size);
-int					t_buff_extend(t_buff *b);
-int					t_buff_add_len(t_buff *b, size_t len);
+size_t				t_buff_extend(t_buff *b);
+size_t				t_buff_add_len(t_buff *b, size_t len);
 
 int					ft_sqrt_int(int nb);
 int					ft_sqrt_ceil(int nb);
